:sparkles: feat: Migrate GamePulse services to Clean Architecture
- Moved files from GamePulse/Services to appropriate layers:
  - Demo.Application for business logic
  - Demo.Infra for infrastructure concerns
- Updated namespaces to reflect new structure
- Added OpenTelemetry package to Demo.Application
- Enhanced code quality with Korean comments and clearer method descriptions
- Resolved build warnings in GamePulse.Test project by utilizing nullable reference types
- Ensured all projects build successfully without warnings
