+++ NET10/Demo.Application/Demo.Application.csproj
@@ -22,6 +22,7 @@
       <PackageReference Include="Serilog.Extensions.Logging" Version="9.0.0" />
       <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.4" />
       <PackageReference Include="FastEndpoints.Security" Version="7.0.1" />
+      <PackageReference Include="OpenTelemetry" Version="1.10.0" />
     </ItemGroup>
 
 </Project>
