+++ NET10/Demo.Infra/Demo.Infra.csproj
@@ -21,6 +21,8 @@
       <PackageReference Include="StackExchange.Redis" Version="2.8.58" />
       <PackageReference Include="FastEndpoints.Security" Version="7.0.1" />
       <PackageReference Include="Bogus" Version="35.6.3" />
+      <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="9.0.7" />
+      <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.4" />
     </ItemGroup>
 
 </Project>
