+++ NET10/docs/clean-architecture-migration/task-001-services-migration.md
@@ -1,0 +1,110 @@
+# Task 001: GamePulse Services를 CleanArchitecture로 마이그레이션
+
+## 개요
+GamePulse/Services 디렉토리의 파일들을 CleanArchitecture 원칙에 따라 Demo.Application과 Demo.Infra 프로젝트로 이동시키는 작업을 수행했습니다.
+
+## 작업 내용
+
+### 1. 파일 분류 및 이동
+
+#### Demo.Application으로 이동 (비즈니스 로직 레이어)
+- `GamePulse/Services/Auth/IAuthService.cs` → `Demo.Application/Services/Auth/IAuthService.cs`
+- `GamePulse/Services/Auth/AuthService.cs` → `Demo.Application/Services/Auth/AuthService.cs`
+
+#### Demo.Infra로 이동 (인프라스트럭처 레이어)
+- `GamePulse/Services/MyTokenService.cs` → `Demo.Infra/Services/MyTokenService.cs`
+- `GamePulse/Services/GamePulseActivitySource.cs` → `Demo.Infra/Services/GamePulseActivitySource.cs`
+
+### 2. 네임스페이스 업데이트
+
+#### Demo.Application
+- `GamePulse.Services.Auth` → `Demo.Application.Services.Auth`
+
+#### Demo.Infra
+- `GamePulse.Services` → `Demo.Infra.Services`
+
+### 3. 의존성 업데이트
+
+#### Demo.Application.csproj
+- OpenTelemetry 패키지 추가: `OpenTelemetry` Version="1.10.0"
+
+#### 참조 업데이트된 파일들
+- `GamePulse/Program.cs`
+- `GamePulse/OpenTelemetryInitialize.cs`
+- `GamePulse/EndPoints/Login/LoginEndpointV1.cs`
+- `GamePulse/Sod/Endpoints/Rtt/RttCommand.cs`
+- `GamePulse/Sod/Endpoints/Rtt/RttEndpointV1.cs`
+- `GamePulse/Sod/Services/SodBackgroundWorker.cs`
+- `GamePulse.Test/Services/AuthServiceTests.cs`
+- `GamePulse.Test/Sod/SodBackgroundWorkerTests.cs`
+
+### 4. 코드 개선사항
+
+#### 한국어 주석 추가
+- 모든 XML 문서 주석을 한국어로 번역
+- 메서드와 클래스에 대한 명확한 설명 추가
+
+#### 코드 품질 향상
+- 일관된 네이밍 규칙 적용
+- 4개 공백 들여쓰기 유지
+- XML 문서 주석 완성
+
+## CleanArchitecture 원칙 적용
+
+### Application Layer (Demo.Application)
+- **IAuthService**: 인증 서비스 인터페이스 (비즈니스 계약)
+- **AuthService**: 인증 비즈니스 로직 구현체
+
+### Infrastructure Layer (Demo.Infra)
+- **MyTokenService**: JWT 토큰 관리 (외부 라이브러리 FastEndpoints 의존)
+- **GamePulseActivitySource**: 텔레메트리 인프라 (OpenTelemetry 의존)
+
+## 검증 사항
+
+### 빌드 검증
+- [x] Demo.Application 프로젝트 빌드 성공
+- [x] Demo.Infra 프로젝트 빌드 성공
+- [x] GamePulse 프로젝트 빌드 성공
+- [x] GamePulse.Test 프로젝트 빌드 성공
+
+### 의존성 검증
+- [x] GamePulse → Demo.Application 참조 확인
+- [x] GamePulse → Demo.Infra 참조 확인
+- [x] Demo.Infra → Demo.Application 참조 확인
+
+### 기능 검증
+- [x] 네임스페이스 참조 업데이트 완료
+- [x] 컴파일 오류 해결
+- [x] 테스트 코드 업데이트 완료
+
+## 다음 단계
+
+1. **빌드 테스트**: 전체 솔루션 빌드 및 테스트 실행
+2. **기능 테스트**: 로그인 엔드포인트 및 JWT 토큰 기능 테스트
+3. **추가 마이그레이션**: 다른 GamePulse 컴포넌트들의 CleanArchitecture 적용 검토
+
+## 파일 변경 요약
+
+### 생성된 파일
+- `Demo.Application/Services/Auth/IAuthService.cs`
+- `Demo.Application/Services/Auth/AuthService.cs`
+- `Demo.Infra/Services/MyTokenService.cs`
+- `Demo.Infra/Services/GamePulseActivitySource.cs`
+
+### 수정된 파일
+- `Demo.Application/Demo.Application.csproj` (OpenTelemetry 패키지 추가)
+- `GamePulse/Program.cs` (네임스페이스 참조 업데이트)
+- `GamePulse/OpenTelemetryInitialize.cs` (네임스페이스 참조 업데이트)
+- `GamePulse/EndPoints/Login/LoginEndpointV1.cs` (네임스페이스 참조 업데이트)
+- `GamePulse/Sod/Endpoints/Rtt/RttCommand.cs` (네임스페이스 참조 업데이트)
+- `GamePulse/Sod/Endpoints/Rtt/RttEndpointV1.cs` (네임스페이스 참조 업데이트)
+- `GamePulse/Sod/Services/SodBackgroundWorker.cs` (네임스페이스 참조 업데이트)
+- `GamePulse.Test/Services/AuthServiceTests.cs` (네임스페이스 참조 업데이트)
+- `GamePulse.Test/Sod/SodBackgroundWorkerTests.cs` (네임스페이스 참조 업데이트)
+
+### 삭제된 파일
+- `GamePulse/Services/` 디렉토리 전체 삭제
+
+## 결론
+
+GamePulse/Services의 모든 파일들이 CleanArchitecture 원칙에 따라 적절한 레이어로 성공적으로 마이그레이션되었습니다. 비즈니스 로직은 Application 레이어로, 인프라스트럭처 관련 코드는 Infrastructure 레이어로 분리되어 의존성 역전 원칙을 준수하게 되었습니다.
