+++ NET10/docs/clean-architecture-migration/task-002-build-warnings-fix.md
@@ -1,0 +1,163 @@
+# Task 002: GamePulse.Test 빌드 경고 수정
+
+## 개요
+GamePulse.Test 프로젝트에서 발생하는 12개의 빌드 경고를 모두 해결했습니다.
+
+## 발생했던 경고들
+
+### 1. CS0219: 사용되지 않는 변수
+- **파일**: `GamePulse.Test/Integration/RttCommandIntegrationTests.cs`
+- **문제**: `expectedRttSeconds` 변수가 선언되었지만 사용되지 않음
+- **해결**: 사용되지 않는 변수 제거
+
+### 2. CS8625: Null 리터럴을 null을 허용하지 않는 참조 형식으로 변환할 수 없음
+- **파일**: `GamePulse.Test/Services/AuthServiceTests.cs`
+- **문제**: `IAuthService.CredentialsAreValidAsync` 메서드의 매개변수가 non-nullable인데 null 값으로 테스트 시도
+- **해결**: 인터페이스와 구현체의 매개변수를 nullable로 변경
+
+### 3. xUnit1012: xUnit Theory에서 null을 non-nullable 타입 매개변수에 사용
+- **파일들**: 
+  - `GamePulse.Test/Services/AuthServiceTests.cs`
+  - `GamePulse.Test/DTO/MyRequestValidatorTests.cs`
+  - `GamePulse.Test/DTO/LoginRequestValidatorTests.cs`
+- **문제**: Theory 메서드의 매개변수가 non-nullable인데 InlineData에서 null 값 사용
+- **해결**: Theory 메서드의 매개변수를 nullable로 변경
+
+### 4. CS8601: 가능한 null 참조 할당
+- **파일들**:
+  - `GamePulse.Test/DTO/LoginRequestValidatorTests.cs`
+  - `GamePulse.Test/DTO/MyRequestValidatorTests.cs`
+- **문제**: nullable 매개변수를 non-nullable 속성에 할당
+- **해결**: null-forgiving operator(`!`) 사용
+
+## 수정된 파일들
+
+### 1. Demo.Application/Services/Auth/IAuthService.cs
+```csharp
+// 변경 전
+Task<bool> CredentialsAreValidAsync(string username, string password, CancellationToken ct);
+
+// 변경 후
+Task<bool> CredentialsAreValidAsync(string? username, string? password, CancellationToken ct);
+```
+
+### 2. Demo.Application/Services/Auth/AuthService.cs
+```csharp
+// 변경 전
+public Task<bool> CredentialsAreValidAsync(string username, string password, CancellationToken ct)
+
+// 변경 후
+public Task<bool> CredentialsAreValidAsync(string? username, string? password, CancellationToken ct)
+```
+
+### 3. GamePulse.Test/Integration/RttCommandIntegrationTests.cs
+```csharp
+// 변경 전
+// Assert
+var expectedRttSeconds = rttMs / 1000.0;
+
+// 로그가 호출되었는지 확인
+
+// 변경 후
+// Assert
+// 로그가 호출되었는지 확인
+```
+
+### 4. GamePulse.Test/Services/AuthServiceTests.cs
+```csharp
+// 변경 전
+public async Task CredentialsAreValidAsync_NullOrEmptyCredentials_OnlyValidWhenPasswordIsAdmin(string username, string password)
+public async Task CredentialsAreValidAsync_DifferentUsernames_WithValidPassword_ReturnsTrue(string username, string password)
+
+// 변경 후
+public async Task CredentialsAreValidAsync_NullOrEmptyCredentials_OnlyValidWhenPasswordIsAdmin(string? username, string? password)
+public async Task CredentialsAreValidAsync_DifferentUsernames_WithValidPassword_ReturnsTrue(string? username, string password)
+```
+
+### 5. GamePulse.Test/DTO/MyRequestValidatorTests.cs
+```csharp
+// 변경 전
+public void Validate_InvalidFirstName_ShouldFail(string firstName, string lastName, int age)
+public void Validate_InvalidLastName_ShouldFail(string firstName, string lastName, int age)
+
+// 변경 후
+public void Validate_InvalidFirstName_ShouldFail(string? firstName, string lastName, int age)
+public void Validate_InvalidLastName_ShouldFail(string firstName, string? lastName, int age)
+
+// 그리고 null-forgiving operator 사용
+FirstName = firstName!,
+LastName = lastName!,
+```
+
+### 6. GamePulse.Test/DTO/LoginRequestValidatorTests.cs
+```csharp
+// 변경 전
+public void Validate_InvalidUsername_ShouldFail(string username, string password)
+public void Validate_InvalidPassword_ShouldFail(string username, string password)
+
+// 변경 후
+public void Validate_InvalidUsername_ShouldFail(string? username, string password)
+public void Validate_InvalidPassword_ShouldFail(string username, string? password)
+
+// 그리고 null-forgiving operator 사용
+Username = username!,
+Password = password!
+```
+
+## 해결 전략
+
+### 1. Nullable Reference Types 활용
+- C# 8.0의 nullable reference types 기능을 활용하여 null 허용 여부를 명시적으로 표현
+- 테스트에서 null 값을 의도적으로 테스트해야 하는 경우 nullable 타입 사용
+
+### 2. Null-forgiving Operator 사용
+- 테스트 코드에서 의도적으로 null 값을 할당해야 하는 경우 `!` 연산자 사용
+- 컴파일러에게 해당 값이 null이 아님을 알려주어 경고 제거
+
+### 3. 사용되지 않는 코드 제거
+- 코드 품질 향상을 위해 사용되지 않는 변수나 코드 제거
+- 테스트 코드의 가독성 향상
+
+## 검증 결과
+
+### 빌드 결과
+- **변경 전**: 12개의 경고 발생
+- **변경 후**: 0개의 경고 (모든 경고 해결)
+
+### 전체 솔루션 빌드
+```
+성공 빌드(8.6초)
+- Demo.Application: 성공
+- Demo.Infra: 성공  
+- Demo.Application.Tests: 성공
+- Demo.Web: 성공
+- GamePulse: 성공
+- Demo.Web.IntegrationTests: 성공
+- Demo.Web.PerformanceTests: 성공
+- GamePulse.Test: 성공 (경고 0개)
+```
+
+## 코드 품질 향상
+
+### 1. 타입 안전성 강화
+- nullable reference types를 통한 null 안전성 향상
+- 컴파일 타임에 null 관련 오류 방지
+
+### 2. 테스트 코드 품질 향상
+- xUnit 분석기 규칙 준수
+- 명확한 테스트 의도 표현
+
+### 3. 코드 일관성 유지
+- 프로젝트 전반에 걸친 일관된 nullable 처리
+- C# 코딩 규칙 준수
+
+## 결론
+
+GamePulse.Test 프로젝트의 모든 빌드 경고를 성공적으로 해결했습니다. 이를 통해:
+
+1. **코드 품질 향상**: 경고 없는 깔끔한 빌드
+2. **타입 안전성 강화**: nullable reference types 활용
+3. **테스트 신뢰성 향상**: 명확한 테스트 의도 표현
+4. **유지보수성 향상**: 일관된 코딩 스타일 적용
+
+모든 프로젝트가 경고 없이 성공적으로 빌드되어 CleanArchitecture 마이그레이션 작업이 안정적으로 완료되었습니다.
